requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
black==22.3.0
boto3==1.22.3
botocore==1.25.3
click==8.1.3
colorama==0.4.4
dnspython==2.2.1
Faker==13.6.0
Flask==2.1.2
Flask-JWT-Extended==4.3.1
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pymongo==4.1.1
pyparsing==3.0.8
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
s3transfer==0.5.2
six==1.16.0
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
